feat: added initial trampoline
trampoline.S assembly blob is now linked to the loader and placed at default
SMBASE on runtime. Please note that the trampoline is not actually doing anything
apart from printing "Crash" to the default console (platform specific as well), and
that the SMI is issued only once on one of the currently running processors.

Signed-off-by: Michal Gorlas <michal.gorlas@9elements.com>