feat: temporarily copying stub code in realmode stage of the boot
Since Linux uses very similar "tool" for moving from 16-bit CPU mode
to the 64-bit mode where the C code can be called, we can reuse this code,
that is, we can place it at default SMBASE + offset (hardcoded to 0x38000 for now)
together with the trampoline used for starting up the kernel.

This commit introduces the initial structure, the code compiles but it is untested at the time being.
Will be used by follow-up commits.

Signed-off-by: Michal Gorlas <michal.gorlas@9elements.com>